refactor: move ReferenceAdapter to new file
