+++ src/language/common/lspLanguageFeatures.ts
@@ -13,7 +13,6 @@ import {
 	CancellationToken
 } from '../../fillers/monaco-editor-core';
 import { fromPosition, toRange, toTextEdit, fromRange } from './CompletionAdapter';
-import { toLocation } from './DefinitionAdapter';
 
 export interface WorkerAccessor<T> {
 	(...more: Uri[]): Promise<T>;
@@ -29,39 +28,7 @@ export * from './DocumentHighlightAdapter';
 
 export * from './DefinitionAdapter';
 
-//#region ReferenceAdapter
-
-export interface ILanguageWorkerWithReferences {
-	findReferences(uri: string, position: lsTypes.Position): Promise<lsTypes.Location[]>;
-}
-
-export class ReferenceAdapter<T extends ILanguageWorkerWithReferences>
-	implements languages.ReferenceProvider
-{
-	constructor(private readonly _worker: WorkerAccessor<T>) {}
-
-	provideReferences(
-		model: editor.IReadOnlyModel,
-		position: Position,
-		context: languages.ReferenceContext,
-		token: CancellationToken
-	): Promise<languages.Location[] | undefined> {
-		const resource = model.uri;
-
-		return this._worker(resource)
-			.then((worker) => {
-				return worker.findReferences(resource.toString(), fromPosition(position));
-			})
-			.then((entries) => {
-				if (!entries) {
-					return;
-				}
-				return entries.map(toLocation);
-			});
-	}
-}
-
-//#endregion
+export * from './ReferenceAdapter';
 
 //#region RenameAdapter
